Put into class
